tests/integration/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
tenacity==8.2.3
docker-compose==1.29.2
nats-py==2.6.0
redis==5.0.1
neo4j==5.14.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.31.0
pydantic==2.5.3
fastapi==0.108.0
allure-pytest==2.13.2
prometheus-client==0.19.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
